refactors order DTOs for post
