--- src/main/java/saboroso/saborosoburguer/DTO/order/ItemFromOrder.java
@@ -1,11 +1,0 @@
-package saboroso.saborosoburguer.DTO.order;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-public interface ItemFromOrder {
-    String identifier();
-    List<String> addOnsIdentifiers();
-    String obs();
-    BigDecimal itemSoldBy();
-}
