+++ src/main/java/saboroso/saborosoburguer/DTO/order/postOrder/DrinkAndQuantityForPostDTO.java
@@ -1,0 +1,6 @@
+package saboroso.saborosoburguer.DTO.order.postOrder;
+
+public record DrinkAndQuantityForPostDTO(
+        String drinkIdentifier,
+        Integer quantity
+) {}
