+++ src/main/java/saboroso/saborosoburguer/DTO/order/postOrder/OrderForPostDTO.java
@@ -1,4 +1,4 @@
-package saboroso.saborosoburguer.DTO.order;
+package saboroso.saborosoburguer.DTO.order.postOrder;
 
 import saboroso.saborosoburguer.models.OrderStatus;
 import saboroso.saborosoburguer.models.PaymentMethod;
@@ -6,15 +6,15 @@ import saboroso.saborosoburguer.models.PaymentMethod;
 import java.math.BigDecimal;
 import java.util.List;
 
-public record OrderDTO(
+public record OrderForPostDTO(
         String orderCode,
         OrderStatus status,
         String clientName,
         String clientPhoneNumber,
         String addressToDeliver,
-        List<BurgerFromOrder> burgers,
-        List<PortionFromOrder> portions,
-        List<DrinkAndQuantity> drinks,
+        List<BurgerForPostOrderDTO> burgers,
+        List<PortionForPostOrderDTO> portions,
+        List<DrinkAndQuantityForPostDTO> drinks,
         PaymentMethod paymentMethod,
         String howClientWillPay,
         BigDecimal totalToPay
